+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Map showing mosque locations
 - Custom Search Results
 - Bg Audio 
+- Subscribe
 
 
 ## [1.0.0] - 2019-05-19